refactor: replace lt function with lessThan method for SemanticVersion comparisons
