+++ src/integration/helm/model/version.ts
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-import {SemanticVersion} from '../base/api/version/semantic-version.js';
+import {SemanticVersion} from '../../../business/utils/semantic-version.js';
 
 /**
  * The response from the helm version command.
@@ -12,13 +12,7 @@ export class Version {
    * Returns a SemanticVersion representation of the version.
    * @returns the helm version
    */
-  public asSemanticVersion(): SemanticVersion {
-    let safeVersion: string = this.version.trim();
-
-    if (safeVersion.startsWith('v')) {
-      safeVersion = safeVersion.slice(1);
-    }
-
-    return SemanticVersion.parse(safeVersion);
+  public asSemanticVersion(): SemanticVersion<string> {
+    return new SemanticVersion<string>(this.version);
   }
 }
